@@ -1,5 +1,6 @@
 plugins {
     id 'com.android.application'
+    id 'org.jetbrains.kotlin.android'
 }
 
 def packageName = "jp.dip.muracoro.comittona"
@@ -15,11 +16,12 @@ android {
     ndkVersion "25.2.9519653"
     defaultConfig {
         applicationId packageName
-        minSdkVersion 9
-        targetSdkVersion 9 // 11以降でSMBまわりでNetworkOnMainThreadExceptionエラー発生
+        minSdkVersion 21
+        // targetSdk11以降でSMBまわりでNetworkOnMainThreadExceptionエラー発生の可能性
+        // targetSdk29はrequestLegacyExternalStorageをセットできる
+        targetSdkVersion 29
         versionCode versionMajor * 10000 + versionMinor * 100 + versionPatch
         versionName "${versionMajor}.${versionMinor}${revision}${versionPatch}"
-        multiDexEnabled true    // メソッド数が65535を超えてもビルド可能にする
         ndk {
             abiFilters "arm64-v8a", "armeabi-v7a", "x86"
             externalNativeBuild {
@@ -64,8 +66,11 @@ android {
     productFlavors {
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_17.toString()
     }
     buildFeatures {
         buildConfig true
@@ -73,23 +78,10 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:multidex:1.0.3'    // メソッド数が65535を超えてもビルド可能にする
-
-    // SDカード書き込み用 Android サポートライブラリ
-    // 古いバージョンを採用することで、minSdkVersionを最小に抑える
-    implementation 'com.android.support:support-compat:24.2.1'
-    implementation 'com.android.support:appcompat-v7:24.2.1'
+    implementation 'androidx.core:core-ktx:1.10.1'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
     // SVGファイル(アイコン表示用)
     implementation 'com.caverock:androidsvg-aar:1.4'
-    // PDF
-    //implementation 'com.joanzapata.pdfview:android-pdfview:1.0.+@aar'
-    // 圧縮ファイル
-    //implementation 'org.apache.commons:commons-compress:1.20'
-    //implementation 'org.tukaani:xz:1.8'
-    // 画像
-    //implementation 'com.github.bumptech.glide:glide:3.6.1'
-
     // SMB2接続用ライブラリ
     implementation 'org.codelibs:jcifs:2.1.35'
-
 }
